requirements.txt
mediapipe==0.10.33
ultralytics==8.4.34
fastapi==0.136.0
uvicorn[standard]==0.43.0
websockets==16.0
python-multipart==0.0.20
openai==2.32.0
opencv-python-headless==4.13.0.92
Pillow==12.2.0
numpy==2.4.4
python-dotenv==1.2.2
pydantic==2.13.1
